requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
google-genai
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
